requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
gunicorn==21.2.0
Pillow==10.1.0
email-validator==2.1.0
marshmallow==3.20.1
Flask-Bcrypt==1.0.1
python-slugify==8.0.1
requests==2.31.0
cloudinary==1.44.1
bleach==6.2.0
pytz==2023.4
asyncio-compat==0.1.2